metcapital/requirements.txt
aiohttp==3.8.3
aiohttp-sse-client
aiosignal==1.2.0
alembic
altair==5.1.1
amqp
anyio
django-cors-headers
argcomplete==1.12.3
asgiref==3.7.2
asttokens==2.0.5
async-generator==1.10
async-timeout
attrs==22.1.0
backoff
base58
billiard
bitarray==2.6.0
black==22.10.0
blinker==1.6.2
boto3
botocore
Brownie==0.5.1
cachetools
celery
certifi==2022.9.24
cffi
cfgv==3.3.1
charset-normalizer==2.1.1
click
click-didyoumean
click-plugins
click-repl
colorlog==6.6.0
contourpy==1.1.0
cryptography
cycler==0.11.0
cytoolz==0.12.0
dataclassy==0.11.1
Django==5.0
djangorestframework==3.14.0
docopt==0.6.2
eip712==0.1.0
et-xmlfile==1.1.0
eth-brownie==1.19.3
eth-event==1.2.3
execnet==1.9.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
fonttools==4.42.1
frozenlist==1.3.1
fsspec==2023.12.2
geckodriver-autoinstaller==0.1.0
gitdb==4.0.10
GitPython==3.1.36
h11==0.14.0
hexbytes==0.2.3
hypothesis==6.27.3
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsoncodable==0.1.7
jsonpickle==3.0.1
jsonschema==3.2.0
k-selenium-cookies==0.0.4
kiwisolver==1.4.5
kproxy==0.0.1
lazy-object-proxy==1.7.1
networkx==3.2.1
nodeenv==1.7.0
noraise==0.0.16
nox==2022.1.7
numpy==1.24.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.2
pathspec==0.10.1
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==2.5.2
playwright==1.36.0
pluggy==1.0.0
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pyarrow==13.0.0
pydeck==0.8.0
pyee==9.0.4
Werkzeug==2.3.4
wrapt==1.14.1
wsproto==1.2.0
xpath-utils==0.0.3
yarg==0.1.9
yarl==1.8.1
zipp==3.16.2
